+++ botx/bot/api/attachments.py
@@ -1,9 +1,9 @@
 import base64
 from typing import Union, cast
 
-from botx.bot.models.commands.enums import AttachmentTypes
-from botx.shared_models.api.enums import (
+from botx.models.enums import (
     APIAttachmentTypes,
+    AttachmentTypes,
     convert_attachment_type_to_domain,
 )
 from botx.shared_models.api_base import VerifiedPayloadBaseModel
