+++ botx/bot/api/status/recipient.py
@@ -2,8 +2,8 @@ from typing import Optional
 from uuid import UUID
 
 from botx.bot.models.status.recipient import StatusRecipient
+from botx.models.enums import APIChatTypes, convert_chat_type_to_domain
 from botx.shared_models.api_base import VerifiedPayloadBaseModel
-from botx.shared_models.chat_types import APIChatTypes, convert_chat_type_to_domain
 
 
 class BotAPIStatusRecipient(VerifiedPayloadBaseModel):
