+++ botx/bot/models/status/recipient.py
@@ -2,8 +2,8 @@ from dataclasses import dataclass
 from typing import Optional
 from uuid import UUID
 
+from botx.models.enums import ChatTypes
 from botx.models.message.incoming_message import IncomingMessage
-from botx.shared_models.chat_types import ChatTypes
 
 
 @dataclass
