+++ botx/client/chats_api/create_chat.py
@@ -7,14 +7,10 @@ from botx.client.chats_api.exceptions import (
     ChatCreationError,
     ChatCreationProhibitedError,
 )
+from botx.models.enums import APIChatTypes, ChatTypes, convert_chat_type_from_domain
 from botx.shared_models.api_base import (
     UnverifiedPayloadBaseModel,
     VerifiedPayloadBaseModel,
-)
-from botx.shared_models.chat_types import (
-    APIChatTypes,
-    ChatTypes,
-    convert_chat_type_from_domain,
 )
 
 try:
