+++ botx/constants.py
@@ -4,3 +4,4 @@ except ImportError:
     from typing_extensions import Final  # type: ignore  # noqa: WPS440
 
 CHUNK_SIZE: Final = 1024 * 1024  # 1Mb
+BOT_API_VERSION: Final = 4
