+++ botx/models/system_events/deleted_from_chat.py
@@ -4,15 +4,14 @@ from uuid import UUID
 
 from pydantic import Field
 
-from botx.bot.api.enums import BotAPICommandTypes
-from botx.bot.models.commands.chat import Chat
 from botx.models.base_command import (
     BotAPIBaseCommand,
     BotAPIChatContext,
     BotCommandBase,
 )
+from botx.models.chat import Chat
+from botx.models.enums import BotAPICommandTypes, convert_chat_type_to_domain
 from botx.shared_models.api_base import VerifiedPayloadBaseModel
-from botx.shared_models.chat_types import convert_chat_type_to_domain
 
 try:
     from typing import Literal
