+++ tests/bot/api/commands/test_incoming_message.py
@@ -19,7 +19,7 @@ from botx import (
     UserEventSender,
     lifespan_wrapper,
 )
-from botx.bot.models.commands.enums import AttachmentTypes
+from botx.models.enums import AttachmentTypes
 from botx.models.message.entities import Forward, Mention, MentionList, Reply
 from botx.shared_models.domain.files import Image
